ui: collect Quote and Truncate helpers
Collect ui formatting helpers in the ui package
