+++ internal/ui/termstatus/status_test.go
@@ -5,7 +5,6 @@ import (
 	"context"
 	"fmt"
 	"io"
-	"strconv"
 	"testing"
 
 	"github.com/restic/restic/internal/terminal"
@@ -56,91 +55,6 @@ func TestSetStatus(t *testing.T) {
 
 	<-term.closed
 	rtest.Equals(t, exp, buf.String())
-}
-
-func TestQuote(t *testing.T) {
-	for _, c := range []struct {
-		in        string
-		needQuote bool
-	}{
-		{"foo.bar/baz", false},
-		{"föó_bàŕ-bãẑ", false},
-		{" foo ", false},
-		{"foo bar", false},
-		{"foo\nbar", true},
-		{"foo\rbar", true},
-		{"foo\abar", true},
-		{"\xff", true},
-		{`c:\foo\bar`, false},
-		// Issue #2260: terminal control characters.
-		{"\x1bm_red_is_beautiful", true},
-	} {
-		if c.needQuote {
-			rtest.Equals(t, strconv.Quote(c.in), Quote(c.in))
-		} else {
-			rtest.Equals(t, c.in, Quote(c.in))
-		}
-	}
-}
-
-func TestTruncate(t *testing.T) {
-	var tests = []struct {
-		input  string
-		width  int
-		output string
-	}{
-		{"", 80, ""},
-		{"", 0, ""},
-		{"", -1, ""},
-		{"foo", 80, "foo"},
-		{"foo", 4, "foo"},
-		{"foo", 3, "foo"},
-		{"foo", 2, "fo"},
-		{"foo", 1, "f"},
-		{"foo", 0, ""},
-		{"foo", -1, ""},
-		{"Löwen", 4, "Löwe"},
-		{"あああああ/data", 7, "あああ"},
-		{"あああああ/data", 10, "あああああ"},
-		{"あああああ/data", 11, "あああああ/"},
-	}
-
-	for _, test := range tests {
-		t.Run("", func(t *testing.T) {
-			out := Truncate(test.input, test.width)
-			if out != test.output {
-				t.Fatalf("wrong output for input %v, width %d: want %q, got %q",
-					test.input, test.width, test.output, out)
-			}
-		})
-	}
-}
-
-func benchmarkTruncate(b *testing.B, s string, w int) {
-	for i := 0; i < b.N; i++ {
-		Truncate(s, w)
-	}
-}
-
-func BenchmarkTruncateASCII(b *testing.B) {
-	s := "This is an ASCII-only status message...\r\n"
-	benchmarkTruncate(b, s, len(s)-1)
-}
-
-func BenchmarkTruncateUnicode(b *testing.B) {
-	s := "Hello World or Καλημέρα κόσμε or こんにちは 世界"
-	w := 0
-	for i := 0; i < len(s); {
-		w++
-		wide, utfsize := wideRune(s[i:])
-		if wide {
-			w++
-		}
-		i += int(utfsize)
-	}
-	b.ResetTimer()
-
-	benchmarkTruncate(b, s, w-1)
 }
 
 func TestSanitizeLines(t *testing.T) {
